Work on AAs. *Updated Alkabor DB is required*
Points, Exp, and Exp % all work great.

You can buy some general AAs up to rank 5. Their effects work both client side and server side.

Any AA on the bottom of the first page and beyond will either not train, is trainable buy buying another ability, or may cause a client desync.

The struct is definitely wrong, but some abilities work on some classes but not others, so it may even be per class :( I am also not sending out a 12 byte packet on opcode 1442 seen on Live. The only column I've identified of it so far is aaid.