Archery and throwing weapons should now work, including animations.
Bolts also work with minor heading issues, but the spellsnew files have to be removed from the client for them to show up. Not authentic to Al'Kabor of course, just something to tinker with.

Made some changes to AAs to try to figure out why activated abilities aren't working properly client side.