DB change required, will be in next Alkabor revision (I'll get it up today.)
Fixed a crash in spawn_conditions caused by NPCs on one way paths.

Added a strict column to spawn_events, to prevent a spawn_condition from firing mid-cycle if need be (will be required by boats.)

Added stuff to get boats working - not finished yet. Boats do not work and there is still debug code left.

Changes to skills to hopefully eventually get guildmasters working correctly.

Removed the +6 tacked on to the z coord during zone-in. EQMac will still continue to hop 5 coords on zone-in. This happened on Alkabor, too. I confirmed it while it was still up.