Fixed merchant price bug.
You will now be stripped of buffs when you zone if your HP is lower than the HP you got from buffs. (Don't worry, if you all hate it I also made it a rule.)

A couple of typo fixes, fiddling with objects to get items within world containers to spawn for a new player.