Corrected the spell scribing related item desyncs.
You will now receive a warning in a case where the server feels you may have desynced (or are about to), but haven't yet desynced client side. This message *should* only occur on the Intel client. But if it does happen on PC client, please report it. Also, I'd welcome discussion on the forums of any patterns that you may see when this message occurs. I have a sneaky suspicion that the longer you are in a zone for, the less likely you are to see this message. Also, I believe the faster you move items around, the more likely you are to see it. For reference, the message is in yellow and is:

"Caution: You may have de-synced. Stopped what you're doing, and log out now to avoid item loss."

Changed some desync logging to make it easier to spot true errors.